requirements.txt
fitz==0.0.1.dev2
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.10
langchain_openai==0.2.2
python-dotenv==1.0.1
streamlit==1.39.0
PyMuPDF==1.24.11
chromadb==0.5.11
protobuf==4.25.5
pysqlite3-binary
rank-bm25==0.2.2
reportlab==4.2.5